+++ documentation/mkdocs.nix
@@ -11,11 +11,10 @@ base.mkDerivation (attrs // {
     mkdir -p $out/share/doc/${name}/${type}
     cp -r site/. $out/share/doc/${name}/${type}
   '';
+  shellCommands = {
+    run = ''mkdocs serve "$@" &'';
+  };
   shellHook = ''
-    preview() {
-      mkdocs serve &
-    }
-
-    echo -e "Use \033[1mpreview\033[0m to look at the docs in a webbrowser, which updates on file save"
+    echo -e "Use \033[1mrun\033[0m to look at the docs in a webbrowser, which updates on file save"
   '';
 })